fix(document): filter out falsy attribute values for inline loaders
to avoid excess script hashes when preemptively register loaders for ISR
(due to different async/defer true/false combinations)

Can be safely reduced by filtering out falsy attribute values before,
as just not setting the attribute means false

refactor:  move code to remove circular dependency

fix(rollup): get rid of rollup warnings

fix(rollup): list all external modules explicitly for each bundle
